Move hash/insert/contains scalar into partial.rs, add more documentation comments and fix primitive test for validation
