+++ vortex-layout/src/layouts/zoned/aggregates/bloom_filter/constant.rs
@@ -2,15 +2,7 @@
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 use vortex_array::arrays::ConstantArray;
-use vortex_array::dtype::DType;
-use vortex_array::dtype::PType;
-use vortex_array::match_each_float_ptype;
-use vortex_array::match_each_integer_ptype;
-use vortex_array::scalar::DecimalValue;
-use vortex_array::scalar::Scalar;
-use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
-use vortex_error::vortex_err;
 
 use crate::layouts::zoned::aggregates::bloom_filter::BloomPartial;
 
@@ -25,92 +17,9 @@ pub(super) fn accumulate_constant(
         return Ok(());
     }
 
-    partial.insert_hash(partial.hash_valid_scalar(scalar)?);
+    partial.insert_valid_scalar(scalar)?;
+
     Ok(())
-}
-
-impl BloomPartial {
-    /// Scalar values must be valid otherwise the function will raise an err.
-    ///
-    /// This function is used by both, for accumulating scalars,
-    /// but also to get a scalar value membership.
-    pub(in crate::layouts::zoned) fn hash_valid_scalar(
-        &self,
-        scalar: &Scalar,
-    ) -> VortexResult<u64> {
-        if scalar.is_null() {
-            return Err(vortex_err!("cannot hash invalid scalars in bloom filter"));
-        }
-
-        Ok(match scalar.dtype() {
-            DType::Extension(_) => {
-                self.hash_valid_scalar(&scalar.as_extension().to_storage_scalar())?
-            }
-            DType::Bool(_) => self.hash(
-                scalar
-                    .as_bool()
-                    .value()
-                    .vortex_expect("non-null boolean value"),
-            ),
-            DType::Primitive(ptype, _) => match ptype {
-                PType::F16 | PType::F32 | PType::F64 => {
-                    match_each_float_ptype!(ptype, |T| {
-                        let value = scalar
-                            .as_primitive()
-                            .typed_value::<T>()
-                            .vortex_expect("non-null primitive value");
-                        self.hash(value.to_bits())
-                    })
-                }
-                _ => match_each_integer_ptype!(ptype, |T| {
-                    let value = scalar
-                        .as_primitive()
-                        .typed_value::<T>()
-                        .vortex_expect("non-null primitive value");
-                    self.hash(value)
-                }),
-            },
-            DType::Decimal(..) => {
-                let decimal = scalar
-                    .as_decimal()
-                    .decimal_value()
-                    .vortex_expect("non-null decimal value");
-                match decimal {
-                    DecimalValue::I8(v) => self.hash(v),
-                    DecimalValue::I16(v) => self.hash(v),
-                    DecimalValue::I32(v) => self.hash(v),
-                    DecimalValue::I64(v) => self.hash(v),
-                    DecimalValue::I128(v) => self.hash(v),
-                    DecimalValue::I256(v) => self.hash(v),
-                }
-            }
-            DType::Utf8(_) => {
-                let buffer = scalar
-                    .as_utf8()
-                    .value()
-                    .vortex_expect("non-null utf8 value");
-                self.hash(buffer.as_bytes())
-            }
-            DType::Binary(_) => {
-                let buffer = scalar
-                    .as_binary()
-                    .value()
-                    .vortex_expect("non-null binary value");
-                self.hash(buffer.as_slice())
-            }
-            other => {
-                return Err(vortex_err!("bloom filter does not support dtype {other}"));
-            }
-        })
-    }
-
-    pub(in crate::layouts::zoned) fn contains_valid_scalar(
-        &self,
-        scalar: &Scalar,
-    ) -> VortexResult<bool> {
-        let hash = self.hash_valid_scalar(scalar)?;
-        Ok(self.find_hash(hash))
-    }
 }
 
 #[cfg(test)]
